% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve the historical price of cryptocurrency}
\usage{
retrieve_data(
  symbol = "BTC-USDT",
  time_period = "1day",
  start_date = "2021-01-01",
  end_date = "2021-01-09"
)
}
\arguments{
\item{symbol}{string of symbol for cryptocurrency}

\item{time_period}{string of time period}

\item{start_date}{string of start date in character}

\item{end_date}{string of end date in character}
}
\value{
dataframe with time and price data.
}
\description{
Retrieve the historical price of cryptocurrency
}
\examples{
\dontrun{retrieve_data("BTC-USDT", "1day", "2021-01-01", "2021-12-31")}
}
