% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_daily_return.R
\name{avg_daily_return}
\alias{avg_daily_return}
\title{Outputs the average daily return of the inputted cryptocurrency price.}
\usage{
avg_daily_return(prices)
}
\arguments{
\item{prices}{A numeric vector representing the inputted cryptocurrency price.}
}
\value{
A numeric value representing the average daily return of the cryptocurrency.
}
\description{
Outputs the average daily return of the inputted cryptocurrency price.
}
\examples{
price_df = data.frame('Date'=c('2022-01-15', '2022-01-16'), 'Close'=c(1.5, 2.5))
avg_daily_return(price_df$Close)
}
