% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_growth_rate.R
\name{daily_growth_rate}
\alias{daily_growth_rate}
\title{Calculate daily growth rate of a cryptocurrency}
\usage{
daily_growth_rate(df, col_name)
}
\arguments{
\item{df}{A Data frame with date and price data.}

\item{col_name}{a string indicate the column holding daily closing price data.}
}
\value{
A dataframe with a new column of daily growth rate.
}
\description{
Calculate daily growth rate of a cryptocurrency
}
\examples{
price_df = data.frame('Date'=c('2022-01-15', '2022-01-16'), 'Close'=c(1.5, 2.5))
daily_growth_rate(price_df, "Close")
}
