% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_price.R
\name{plot_price}
\alias{plot_price}
\title{Plot the price of cryptocurrency}
\usage{
plot_price(df = price_df)
}
\arguments{
\item{price_df}{data frame with time and price data.}
}
\value{
ggplot object
}
\description{
Plot the price of cryptocurrency
}
\examples{
price_df <- data.frame('Symbol'=c('BTC-USDT','BTC-USDT'), 'Date'=c('2022-01-15', '2022-01-16'), 'Close'=c(1.5, 2.5))
\dontrun{plot_price(df = price_df)}
}
