% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_price.R
\name{plot_price}
\alias{plot_price}
\title{Plot the price of cryptocurrency}
\usage{
plot_price(price_df, name, window)
}
\arguments{
\item{price_df}{data frame with time and price data.}

\item{name}{character name of the cryptocurrency of interest}

\item{window}{vector of two dates in character}
}
\value{
ggplot object
}
\description{
Plot the price of cryptocurrency
}
\examples{
plot_price(price_df, "BTCBitcoin", ["2021-01-01", "2021-12-31"])
}
